tirgger function done

+++ settingWindow.h
@@ -38,7 +38,7 @@ private:
     TriggerFunctions *_triggerSetting;
 
     //double to int converter
-    double doubleToQint8(double value, GlobalEnumatedAndExtern::eRangeValue range);
+    quint8 doubleToQint8(double value, GlobalEnumatedAndExtern::eRangeValue range);
 
 public slots:
     void _recievedNbFrameSavedChanged(quint64 nbFrameChanged);
